# Optimize vector x single-scalar multiplication over GF(2^8), using AVX2 and SSSE3 intrinsics