--- a/Cargo.toml
+++ b/Cargo.toml
@@ -44,11 +44,13 @@ harness = false
 [profile.optimized]
 inherits = "release"
 codegen-units = 1
-lto = "thin"
+lto = true
 panic = "abort"
+opt-level = 3
 
 [profile.test-release]
 inherits = "release"
 debug = true
 debug-assertions = true
 overflow-checks = true
+opt-level = 3