--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rlnc"
-version = "0.5.0"
+version = "0.6.0"
 edition = "2024"
 resolver = "3"
 rust-version = "1.85.0"