--- a/Makefile
+++ b/Makefile
@@ -18,8 +18,8 @@ format: ## Formats source tree
 
 .PHONY: test
 test: ## Run all tests
-	$(BACKTRACE) cargo test --profile test-release
-	$(BACKTRACE) cargo test --profile test-release --features parallel
+	$(BACKTRACE) RUSTFLAGS="-C target-cpu=native" cargo test --profile test-release
+	$(BACKTRACE) RUSTFLAGS="-C target-cpu=native" cargo test --profile test-release --features parallel
 
 .PHONY: test-wasm
 test-wasm: ## Run all tests in WASM environment
@@ -31,8 +31,8 @@ coverage: ## Generates HTML code coverage report, using `cargo-tarpaulin`
 
 .PHONY: bench
 bench: ## Run all benchmarks
-	cargo bench --profile optimized
-	cargo bench --profile optimized --features parallel
+	RUSTFLAGS="-C target-cpu=native" cargo bench --profile optimized
+	RUSTFLAGS="-C target-cpu=native" cargo bench --profile optimized --features parallel
 
 .PHONY: clean
 clean: ## Removes cargo target directory
@@ -40,6 +40,6 @@ clean: ## Removes cargo target directory
 
 .PHONY: example
 example: ## Runs the Full RLNC example program
-	cargo run --example full_rlnc
-	cargo run --example full_rlnc --features parallel
+	RUSTFLAGS="-C target-cpu=native" cargo run --example full_rlnc
+	RUSTFLAGS="-C target-cpu=native" cargo run --example full_rlnc --features parallel
 	cargo run --example full_rlnc --target wasm32-wasip1 --no-default-features