--- a/src/common/gf256.rs
+++ b/src/common/gf256.rs
@@ -1,17 +1,25 @@
 //! Following GF(2**8) logarithm and exponentiation tables are generated using
 //! Python script @ https://gist.github.com/itzmeanjan/0b2ec3f378de2c2e911bd4bb5505d45a.
 
-use crate::common::macros::{generate_gf256_simd_mul_row, generate_gf256_simd_mul_table};
 use rand::Rng;
 use rand::distr::{Distribution, StandardUniform};
+use std::ops::{Add, AddAssign, Div, Mul, Neg, Sub};
+
+#[cfg(not(feature = "parallel"))]
+use crate::common::macros::{generate_gf256_simd_mul_row, generate_gf256_simd_mul_table};
+
+#[cfg(not(feature = "parallel"))]
 use std::arch::x86_64::{
     _mm_and_si128, _mm_lddqu_si128, _mm_set1_epi8, _mm_shuffle_epi8, _mm_srli_epi64, _mm_storeu_si128, _mm_xor_si128, _mm256_and_si256, _mm256_lddqu_si256,
     _mm256_set1_epi8, _mm256_shuffle_epi8, _mm256_srli_epi64, _mm256_storeu_si256, _mm256_xor_si256,
 };
-use std::ops::{Add, AddAssign, Div, Mul, Neg, Sub};
 
-const GF256_BIT_WIDTH: usize = u8::BITS as usize;
 const GF256_ORDER: usize = u8::MAX as usize + 1;
+
+#[cfg(not(feature = "parallel"))]
+const GF256_BIT_WIDTH: usize = u8::BITS as usize;
+
+#[cfg(not(feature = "parallel"))]
 const GF256_HALF_ORDER: usize = 1usize << (GF256_BIT_WIDTH / 2);
 
 const GF256_LOG_TABLE: [u8; GF256_ORDER] = [
@@ -47,13 +55,16 @@ const GF256_EXP_TABLE: [u8; 2 * GF256_ORDER - 2] = [
 /// AVX2 and SSSE3 optimized SIMD multiplication over GF(2^8) uses this lookup table, which is generated following
 /// https://github.com/ceph/gf-complete/blob/a6862d10c9db467148f20eef2c6445ac9afd94d8/src/gf_w8.c#L1100-L1105.
 /// This table holds `htd->low` part, described in above link.
+#[cfg(not(feature = "parallel"))]
 const GF256_SIMD_MUL_TABLE_LOW: [[u8; 2 * GF256_HALF_ORDER]; GF256_ORDER] = generate_gf256_simd_mul_table!(true);
 
 /// AVX2 and SSSE3 optimized SIMD multiplication over GF(2^8) uses this lookup table, which is generated following
 /// https://github.com/ceph/gf-complete/blob/a6862d10c9db467148f20eef2c6445ac9afd94d8/src/gf_w8.c#L1100-L1105.
 /// This table holds `htd->high` part, described in above link.
+#[cfg(not(feature = "parallel"))]
 const GF256_SIMD_MUL_TABLE_HIGH: [[u8; 2 * GF256_HALF_ORDER]; GF256_ORDER] = generate_gf256_simd_mul_table!(false);
 
+#[cfg(not(feature = "parallel"))]
 fn gf256_inplace_mul_vec_by_scalar(vec: &mut [u8], scalar: u8) {
     if vec.is_empty() {
         return;
@@ -135,6 +146,7 @@ fn gf256_inplace_mul_vec_by_scalar(vec: &mut [u8], scalar: u8) {
 /// to enjoy full benefits of compiler optimization.
 ///
 /// I originally discovered this technique in https://www.snia.org/sites/default/files/files2/files2/SDC2013/presentations/NewThinking/EthanMiller_Screaming_Fast_Galois_Field%20Arithmetic_SIMD%20Instructions.pdf.
+#[cfg(not(feature = "parallel"))]
 pub fn gf256_mul_vec_by_scalar(vec: &[u8], scalar: u8) -> Vec<u8> {
     let mut result = vec.to_vec();
     gf256_inplace_mul_vec_by_scalar(&mut result, scalar);
@@ -151,6 +163,7 @@ pub fn gf256_mul_vec_by_scalar(vec: &[u8], scalar: u8) -> Vec<u8> {
 ///
 /// You have to compile with `RUSTFLAGS="-C target-cpu=native -C target-feature=+avx2,+ssse3"`
 /// flag to hint the compiler so that it generates best code.
+#[cfg(not(feature = "parallel"))]
 pub fn gf256_inplace_add_vectors(vec_dst: &mut [u8], vec_src: &[u8]) {
     if cfg!(target_arch = "x86_64") && is_x86_feature_detected!("avx2") {
         unsafe {