--- a/src/common/macros.rs
+++ b/src/common/macros.rs
@@ -30,6 +30,7 @@
 //!     .collect::<Vec<Vec<u8>>>();
 //! ```
 
+#[cfg(not(feature = "parallel"))]
 macro_rules! generate_gf256_simd_mul_row {
     ($a:expr,$is_low_part:expr) => {
         [
@@ -74,6 +75,7 @@ macro_rules! generate_gf256_simd_mul_row {
     };
 }
 
+#[cfg(not(feature = "parallel"))]
 macro_rules! generate_gf256_simd_mul_table {
     ($is_low_part:expr) => {
         [
@@ -337,5 +339,8 @@ macro_rules! generate_gf256_simd_mul_table {
     };
 }
 
+#[cfg(not(feature = "parallel"))]
 pub(crate) use generate_gf256_simd_mul_row;
+
+#[cfg(not(feature = "parallel"))]
 pub(crate) use generate_gf256_simd_mul_table;