--- a/src/common/mod.rs
+++ b/src/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod errors;
 pub mod gf256;
+pub mod macros;