Discovered a technique for fast vector x single-scalar multiplication over $GF(2^8)$, which can be used in RLNC encoding part. With this optimization implemented, single-threaded RLNC erasure-coding speed just became ~(3-4)x of what it was in https://github.com/itzmeanjan/rlnc/releases/tag/v0.5.0. Recoding speed also enjoys similar boost, as Recoding used encoding under-the-hood.

> [!NOTE]
> The technique is described in https://www.snia.org/sites/default/files/files2/files2/SDC2013/presentations/NewThinking/EthanMiller_Screaming_Fast_Galois_Field%20Arithmetic_SIMD%20Instructions.pdf.
> Also found a C implementation of this optimization technique in https://github.com/ceph/gf-complete/blob/a6862d10c9db467148f20eef2c6445ac9afd94d8/src/gf_w8.c#L1029-L1037.

Metric | In 469d3f8265 commit (v0.5.0 tag) of `rlnc`  crate | In acc1d4bb commit of `rlnc` crate | Change
--- | --- | --- | ---
Latency to produce a single erasure-coded share from 16MB input blob, split into 16 original shares | 8.794ms  | 2.572ms | -70%
Throughput of producing erasure-coded shares, from 16MB original blob, split into 16 original shares | 1.887GB/s | 6.453GB/s |  +241%